Time: 53 ms (63.46%), Space: 10.4 MB (75.64%) - LeetHub

+++ 629-k-inverse-pairs-array/629-k-inverse-pairs-array.cpp
@@ -11,7 +11,7 @@ public:
                 if(j>=i){
                     x-=v[i-1][j-i];
                 }
-                v[i][j] = (v[i][j]+x)%mod;
+                v[i][j] = ((long long)v[i][j]+x)%mod;
             }
         }
         return v[n][k];
